feat(prescriptions): add patient order PDF generation
This commit introduces the functionality of PDF generation for patient orders in the ClinicManagement::PrescriptionsController. A new action, `generate_order_pdf`, has been added which takes an `order_number` parameter and responds with an appropriate PDF.

Additionally, changes have been made to the associated views to include a form for providing the `order_number` and a list of links for generating these PDFs. The routing configuration has also been updated to handle the new `generate_order_pdf` route.

The 'set_appointment' filter in PrescriptionsController now excludes the 'generate_order_pdf' action. This is necessary because this action doesn't require an appointment to be set beforehand. Consequently, the absence of an appointment will not interfere with the generation of patient order PDFs.