feat: Exclude leads with attendance within the last year in leads_controller
The code changes in leads_controller.rb include excluding leads that have had attendance within the last year. This is achieved by retrieving the IDs of leads with attendance as true within the last year and excluding them from the query results. The commit message suggests adding this exclusion logic to improve the functionality of the leads_controller.